gateway/pyproject.toml
[project]
name = "gateway"
version = "0.1.0"
description = "Gateway service for ai-manus runtime"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.121.2",
    "httpx>=0.28.1",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.12.0",
    "pyjwt[crypto]>=2.8.0",
    "redis>=5.0.1",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.0",
]
